linux: Refresh devices after resume from sleep
When we wake up from suspend, the battery levels might have changed
drastically. Since we poll batteries it can take quite some time
until the new level is picked up by the polling code.
We now listen to the "PrepareForSleep" signal emitted by logind,
that will get called on suspend and resume and in the latter case
we manually trigger a refresh of devices, to get the new battery
levels immediately after resume.

https://bugs.freedesktop.org/show_bug.cgi?id=99763